Use size_t for array list length and size

+++ arraylist.h
@@ -23,8 +23,8 @@ typedef void (array_list_free_fn) (void *data);
 struct array_list
 {
   void **array;
-  int length;
-  int size;
+  size_t length;
+  size_t size;
   array_list_free_fn *free_fn;
 };
 
@@ -35,15 +35,15 @@ extern void
 array_list_free(struct array_list *al);
 
 extern void*
-array_list_get_idx(struct array_list *al, int i);
+array_list_get_idx(struct array_list *al, size_t i);
 
 extern int
-array_list_put_idx(struct array_list *al, int i, void *data);
+array_list_put_idx(struct array_list *al, size_t i, void *data);
 
 extern int
 array_list_add(struct array_list *al, void *data);
 
-extern int
+extern size_t
 array_list_length(struct array_list *al);
 
 extern void
